Continue Refactoring
* Split the matrix functions into two files, `dense_matrix.cpp` and `sparse_matrix.cpp`
* Refactor tests completely to align with new API
